feat(agent-tars): refine server code structure
