+++ multimodal/agent-tars-server/src/api/index.ts
@@ -1,0 +1,20 @@
+import express from 'express';
+import cors from 'cors';
+import { registerAllRoutes } from './routes';
+import { ServerOptions } from '../models/ServerOptions';
+
+/**
+ * Setup API middleware and routes
+ * @param app Express application instance
+ * @param options Server options
+ */
+export function setupAPI(app: express.Application, options: ServerOptions) {
+  // Apply CORS middleware
+  app.use(cors(options.getEffectiveCorsOptions()));
+
+  // Apply JSON body parser middleware
+  app.use(express.json());
+
+  // Register all API routes
+  registerAllRoutes(app);
+}
