--- multimodal/agent-tars-server/src/cli/index.ts
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) 2025 Bytedance, Inc. and its affiliates.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './storage-info';
