+++ multimodal/agent-tars-server/src/event-stream.ts
