+++ multimodal/agent-tars-server/src/models/index.ts
@@ -1,0 +1,2 @@
+export { default as AgentSession } from './AgentSession';
+export { default as ServerOptions } from './ServerOptions';
